requirements.txt
# Python dependencies for foundry-bootstrap orchestrator
pyyaml>=6.0
click>=8.0
rich>=13.0 